requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Statistical modeling
statsmodels>=0.14.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Google Sheets integration
gspread>=5.10.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
